server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.6",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
]
